backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Face Recognition
facenet-pytorch==2.5.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# Vector DB
qdrant-client==1.11.0

# Graph DB
neo4j==5.23.0

# LLM & AI
groq==0.9.0
langchain==0.2.14
langchain-groq==0.1.9
sentence-transformers==3.0.1

# Utilities
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.0
numpy>=1.24.0
